# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: CyCove Backend API
  version: 0.6.0-draft
  description: >
    HTTP API contract for the CyCove relay backend, formalizing the surface
    sketched in the vault note (Projects/CyCove.md, Architecture -> Backend).
    The server is a deliberate "dumb pipe": it never sees plaintext message
    content and never holds long-term encryption keys. Realtime delivery is
    covered separately in asyncapi.yaml.
    STATUS (2026-07-19): reshaped around matrix-sdk-crypto-wasm's actual
    protocol requirements — see docs/crypto-integration-notes.md for why.
    /keys/upload, /keys/query, /keys/claim replace the original /prekeys
    endpoints (Matrix-shaped batch operations, not this API's own simpler
    shapes). user_id and device_id are now client-generated, not
    server-assigned — required because the client's OlmMachine bakes both
    into a self-signed device_keys blob before /register is ever called.
    STATUS (2026-07-19, later same day): added a real username directory
    (/users/*), rate limiting on the most abuse-prone endpoints, and opaque
    encrypted contact backup (/account-data/*) — see
    docs/crypto-integration-notes.md and THREAT_MODEL.md for the disclosures
    that come with both.
    STATUS (2026-07-20): added /login (re-authenticate an existing account on
    the device that created it) and wired the web client to the previously
    backend-only /devices/pairing-token and /devices/link (real "linked
...
    more use for it.
...
servers:
...
paths:
...
  /devices/pairing-token:
    post:
      tags: [devices]
      summary: Issue a short-lived, single-use token for linking a new device
      description: >
        Called by an already-logged-in device from the web client's "Link a
        device" action; the resulting pairing code (this token bundled with
        the caller's own user_id, encoded client-side — see
        web/src/lib/store.ts's encodePairingCode) is what gets pasted into a
        second device's "Link this device" screen, which calls
        POST /devices/link below. See docs/ux-flows.md.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Pairing token issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  pairing_token:
                    type: string
                  expires_in:
                    type: integer
                    description: Seconds until the token expires (300).
        '401':
          $ref: '#/components/responses/Unauthorized'
  /devices/link:
    post:
      tags: [devices]
      summary: Link a new device to an existing account via QR/pairing code
      description: >
        Called by a new device after an already-verified device generated a
        pairing token (POST /devices/pairing-token) and the new device
        scanned/entered it. The token is single-use. device_id and
        device_keys follow the same client-generated, self-signed pattern as
        POST /register — see docs/crypto-integration-notes.md.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pairing_token, device_id, device_keys]
              properties:
                pairing_token:
                  type: string
                  description: Short-lived token issued by POST /devices/pairing-token.
                device_id:
                  type: string
                device_keys:
                  type: object
                one_time_keys:
                  type: object
                push_token:
                  type: string
                  nullable: true
      responses:
        '201':
          description: Device linked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: Pairing token not found or expired
        '409':
          description: device_id already taken
  /devices:
    get:
      tags: [devices]
      summary: List every device on the caller's own account
      description: >
        Added alongside QR-code linking (docs/ux-flows.md): a QR code can be
        captured by a camera from across a room in an instant, unlike a
        copy-pasted string, so an easy way to see and undo an unwanted link
        matters more once QR exists. Returns the caller's own device too —
        the client marks which one that is by comparing IDs it already has.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: The account's devices
          content:
            application/json:
              schema:
                type: object
                properties:
                  devices:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        createdAt:
                          type: string
                          format: date-time
                        lastSeenAt:
                          type: string
                          format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'
  /devices/{deviceId}:
    delete:
      tags: [devices]
      summary: Revoke a device — ends its sessions and removes it
      description: >
        Only works on a device belonging to the caller's own account (checked
        server-side, not just left to client trust — otherwise any
        authenticated user could revoke any device by guessing IDs).
        Invalidates every live session token for the target device
        immediately (see backend/src/lib/auth.ts's revokeDeviceSessions),
        then deletes its row — OneTimeKey and ToDeviceQueue rows for it
        cascade automatically. Revoking the caller's own current device is
        allowed, not blocked; it ends that session too, same as logging out.
      security:
        - bearerAuth: []
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Revoked
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No such device on the caller's own account (also returned, not a 403, if the device belongs to someone else — doesn't confirm the ID exists on another account)
...
  /contacts/request:
    post:
      tags: [contacts]
      summary: Notify another account of a contact request
      description: >
        Deliberately a plain HTTP action, not a WS to-device send — the
        content this carries (the requester's username, already resolvable
        via /users/by-username anyway) needs to be server-visible for
        rate-limiting to target contact-request spam specifically, which
        isn't possible against opaque Olm ciphertext. See
        docs/crypto-integration-notes.md. Delivered via the same
        relayToDevice() used by WS sendToDevice and POST /messages, so it
        behaves identically whether the recipient is online or not.
        contact_accept/contact_decline stay on the WS to-device path (see
        asyncapi.yaml) as plain, non-Olm-encrypted content, for the same
        reason as this endpoint.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [recipient_device_id]
              properties:
                recipient_user_id:
                  type: string
                recipient_device_id:
                  type: string
                username:
                  type: string
                  nullable: true
                  description: The requester's own self-chosen username, if set — lets the recipient show something before ever seeing a share code.
      responses:
        '202':
          description: Queued/delivered for the recipient
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /account-data/{dataType}:
    put:
      tags: [account-data]
      summary: Upload an encrypted backup blob (contacts or cross-signing keys)
      description: >
        Opaque ciphertext — the server never sees plaintext, only
        stores/serves what it's given, same "dumb pipe" pattern as
        device_keys/one_time_keys/to_device_queue. Encrypted client-side
        with a key derived from the recovery key (WebCrypto PBKDF2 +
        AES-GCM — not matrix-sdk-crypto-wasm's key-backup feature, which is
        Megolm-room-key-specific and doesn't fit here; see
        docs/crypto-integration-notes.md). Message content is deliberately
        never backed up this way — see THREAT_MODEL.md and
        Projects/CyCove.md. `dataType` is `contacts` (original use) or
        `cross-signing` (the private cross-signing key export, restored onto
        a newly linked device that was given the recovery key) — one row per
        [user, dataType] (backend/prisma/schema.prisma's EncryptedBackup).
      security:
        - bearerAuth: []
      parameters:
        - name: dataType
          in: path
          required: true
          schema:
            type: string
            enum: [contacts, cross-signing]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ciphertext, iv]
              properties:
                ciphertext:
                  type: string
                  description: Base64-encoded AES-GCM ciphertext.
                iv:
                  type: string
                  description: Base64-encoded initialization vector, fresh per encryption.
      responses:
        '204':
          description: Stored (upserted)
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    get:
      tags: [account-data]
      summary: Fetch the caller's encrypted backup blob, if any
      security:
        - bearerAuth: []
      parameters:
        - name: dataType
          in: path
          required: true
          schema:
            type: string
            enum: [contacts, cross-signing]
      responses:
        '200':
          description: Backup found
          content:
            application/json:
              schema:
                type: object
                properties:
                  ciphertext:
                    type: string
                  iv:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No backup uploaded yet
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Device-scoped session token, issued by /register or /devices/link.
        Implemented as an opaque random token mapped to a device ID in Redis
        (backend/src/lib/auth.ts) — no expiry or refresh yet, that's still an
        open design question, not just an unimplemented one.
  schemas:
    RegistrationResult:
      type: object
      description: Shared response shape for POST /register and POST /devices/link.
      properties:
        user_id:
          type: string
          description: Only present on POST /register — POST /devices/link doesn't create a new user.
        device_id:
          type: string
        session_token:
          type: string
          description: Bearer token for this device's session — pass as `Authorization Bearer <token>` on authenticated routes.
        one_time_key_counts:
          type: object
          description: Same shape as /keys/upload's response — the client needs this to satisfy markRequestAsSent after its initial key upload.
          properties:
            signed_curve25519:
              type: integer
  responses:
    BadRequest:
      description: Malformed request
    Unauthorized:
      description: Missing or invalid session token
    RateLimited:
      description: >
        Rate limit exceeded. First-pass limits shipped 2026-07-19
        (backend/src/lib/rateLimit.ts, fixed-window, per-device or per-IP
        depending on the endpoint) — not yet tuned against real usage, see
        THREAT_MODEL.md -> Open items.